+++ app/Filament/Resources/LeaseResource/States/LeaseOptionState.php
@@ -27,7 +27,12 @@ final class LeaseOptionState extends LeaseState
      */
     public function transitionToCancelled(CancellationDataObject $cancellationDataObject): bool
     {
-        return DB::transaction(fn(): bool => $this->lease->setStatus(LeaseStatus::Cancelled)->registerCancellation($cancellationDataObject->getReason()));
+        return DB::transaction(function () use ($cancellationDataObject): bool {
+            $status = LeaseStatus::Cancelled;
+            $auditMessage = trans('Heeft de status van een verhuring gewijzigd naar :status', ['status' => $status->getLabel()]);
+
+            return $this->lease->setStatus(newStatus: $status, auditMessage: $auditMessage)->registerCancellation($cancellationDataObject->getReason());
+        });
     }
 
     /**